+++ src/app/modules/shared/public-api.ts
@@ -1,0 +1,10 @@
+export {
+  ConfigurationService,
+  IConfiguration,
+} from './services/configuration/configuration.service';
+export {
+  IMusicLibrary,
+  IMusicLibraryFolder,
+  IMusicLibraryItem,
+  MusicLibraryService,
+} from './services/music-library/music-library.service';