+++ src/app/modules/shared/shared.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ClarityModule } from '@clr/angular';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+  ],
+  exports: [
+    ClarityModule,
+  ]
+})
+export class SharedModule { }