+++ CHANGELOG.md
@@ -5,3 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Can add a folder to the music library
+- Can list all files in the music library
+- Use clarity design